Improve boundary integration test and add package metadata test
