+++ tests/integration/test_solver_performance.py
