+++ tests/unit/tofea/test_solvers.py
